packages/fastapi-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-openai>=0.0.2
langgraph==0.0.20
pydantic>=2.5.0
python-dotenv==1.0.0
sse-starlette==1.8.2
httpx==0.25.2
tavily-python
composio-core==0.3.0
